Move gene component into separate react component
